+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## 1.0.1 under development
 
+- Bug #96: Fix unexpected redirects from `Locale` middleware on GET requests (@vjik)
+- Bug #97: Don't search locale in cookies when `$cookieDuration` is null (@vjik)
 - Chg #95: Remove unused network utilities dependency (@arogachev)
 
 ## 1.0.0 May 22, 2023